[viewer] Add forms with params and move forms to its own forms.py file
In order to fill forms data depending if a param is defined or not,
we need to add support for receiving params in the form classes.

For example, if we want the projects for an ecosystem, we just pass
the eco_name param to the ProjectsForm class and only the projects
for this ecosystem will be shown. The basic code for do that is:

class ProjectsForm(forms.Form):
    widget = forms.Select(attrs={'size': SELECT_LINES, 'class': 'form-control'})

    def __init__(self, *args, **kwargs):
        self.eco_name = None
        if 'eco_name' in kwargs:
            self.eco_name = kwargs.pop('eco_name')
        super(ProjectsForm, self).__init__(*args, **kwargs)