[editor] Massive refactoring to allow the reuse of code between web pages
Durig the development of the new status page, in order to reuse the current
forms logic as much as possible, the select methods like select_ecosystem or
select_project has been refactored to be generic and reusable in
different scenarios.

The methods that are shared between web pages are now at the top of
views.py and the logic for each web page is grouped now and separated
by commnets.