Add robustness to web UI ignoring ecosystem forms with empty name
This patch aims to fix a side-effect that appeared when the
`onclick` event was added to the select ecosystem forms.
As the first option is empty (to show the user that no ecosystem
has been selected), when that empty option is clicked it submits
the data and it provoked an error. Now, out `select_ecosystem`
method ignores forms with an empty name, returning to the editor
keeping its previous state.