[projects] Make all URLs relative in project app
In this way, if the name of the applicacion changes in the future,
no change is needed in the URLs.

The "/" for Bestiary is now redirected to "/projects". It is not
possible anymore to use projects from "/". In this way, we start to
isolate the applicacions in its own directories.