Add information messages to web UI related to important actions
Using a block from `base` template, `EditorState` class has been
modified so now it includes the variable `msg` to display that
string into an alert on the top of the page.

Fix `add_project` method: if there is a project that belongs to
an ecosystem which is removed, that project can be added "as is"
to a new ecosystem, printing a warning message.

Unify `formaction` strings with double quotes and `Save` and
`Remove` button colors, `Edit` ecosystem button size.