[models] Add Transaction and Operation classes
`Transaction` class is meant to represent objects which are
created inside a context (e.g., a method from the API)
where the database is going to be modified in some way.
The `name` parameter is a choice field which is related with
the name of the method creating the transaction.

`Operation` class is meant to represent atomic interactions
with the database (either add, delete or update something)
and the entities which are involved in this interaction.
An `Operation` object is including the original arguments
from the method creating the operation.

Note that the type of allowed operations from `Operation`
class are defined in an inner *enum* class named `OpTypeChoice`.

As `Operation` model is using the JSONField from `django_mysql`
module, it is now added to the `requirements.txt` file as a new
dependency.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>