[log] Add TransactionsLog class
This new class creates new `TransactionsLog` objects, generating
a new `Transaction` object with each instance. The class contains
a method to create `Operation` objects from the generated
transaction, and it also contains a `close` method, setting a
`closed_at` date to that Transaction object and also
a boolean flag.

A new error has been added too, `ClosedTransactionError`, which
should be raised when trying to log an operation over a closed
transaction.

The `BestiaryContext` object stores a context where the
transaction is created. It contains the username from the user
(as defined in Django) who called the method causing the
transaction.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>