[schema] Add queries for transactions and operations
These new queries allow to export the list of transactions
and operations from the database with the following format:

query {
    transactions {
      name
      createdAt
      tuid
      isClosed
      closedAt
    }
}

query {
    operations {
      ouid
      opType
      entityType
      target
      timestamp
      args
      trx {
        name
        createdAt
      }
    }
}

Both queries support filtering for most of the fields from
their corresponding models, including `from_date` and `to_date`
filters for both models (referencing to `created_at` field
for transactions and `timestamp` for operations). This date
can be passed as filter as a string following ISO 8601 format.

As the `args` field from `Operations` model is encoded in a
serialized JSON, it was necessary to define a explicit conversion
to return these encoded JSON fields as dictionaries. The class
in charge of this process is named `OperationArgsType` and it
inherits from `GenericScalar` class, which is a basic type defined
in `graphene` module representing a generic GraphQL scalar value
that could be String, Boolean, Int, Float, List or Object.

Note: These queries returns paginated results. When performing a
query asking for paginated results, there are two parameters:
- `pageSize`: Number of elements we want to obtain per page.
 If `pageSize` value is not set, the value will be set using
 a configuration parameter defined in Django settings, named
 `DEFAULT_GRAPHQL_PAGE_SIZE`.
- `page`: Number of the page we are asking for.
 If `page` value is not set, it will return the first page
 by default.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>