[schema] Complete basic schema structure
In `graphene` module, the schema is built with the classes where
the queries and the mutations are defined. So far, only the
authentication mutations have been added, but no queries.
After adding the queries for returning `Transaction` and
`Operation` objects, the basic schema structure can be formed.

This change includes the addition on the GraphQL endpoint in
the `urls.py` script and the definition of the graphene
`schema` object.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>