[models] Add Ecosystem model and inital migrations file
Ecosystem class is meant to represent an abstract set of projects
which may share a common context.

Every ecosystem object must have a name `name` (which is going to be
unique across the database) and a unique `id`, which by default is generated
automatically setting an incremental number. Optionally, it may have a
title (`title`) and a `description` field containing a brief text describing
what the ecosystem is about.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>