[models] Add Project model and inital migrations file
Projects are defined as a set of data locations which have to be
grouped under the same entity. Moreover, this grouping may have a
hierarchy by defining n sub-projects.

Every project object must have a name (`name`) and must belong to
one single Ecosystem (`ecosystem`).
Optionally, it may have a title (`title`), and a relation with
a parent project (`parent_project`).

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>