[models] Remove unnecessary field in Ecosystem and Project models
Both models contained the `meta` field `unique_together` with
one model field as value (`name`). This is not needed because
this makes sense when more than one field is considered for
having unique tuples.

Moreover, this was causing to get a MySQL warning about duplicate
indexes.

To remove this warning, the migrations were generated from scratch.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>