[db] Link a project object with another
This method sets the `parent_project` field value from `project`
with the `Project` object from `parent_project`. If this field
already had a value, it will be overwritten. In case the parent
project is invalid, the method will raise an `ValueError`.

The reasons for considering a parent as invalid are:
  - The parent is already set to the project.
  - The parent and the project are the same.
  - The parent belongs to a different ecosystem.
  - If the project is not a root one, when the parent comes from
   a different root project.
  - The parent is a descendant from the project.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>