[api] Move a project to a parent project
Link a project with another one with a child-parent relation.
If a project is not linked to any project, it is a root project.
Once the project is  linked with another one, it won't be a root
project anymore.

If the source project was already linked to a parent project, it
will be overwritten.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>