[CI] Migrate from Travis CI to GitHub Actions
Travis CI has brought a few changes in its plans
which is causing friction to the current workflow.

This commit removes the support of Travis CI and adds
the support of GitHub Actions as the new CI service.

The grimoire-elk package installation is also updated
in the requirements.txt file.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>