[ui] Create ProjectForm component
Adds a 'ProjectForm' component that can be used
to add a project to a given ecosystem.
The form suggests a name for the project based on its
title in lowercase, replacing spaces with '-'.
When the select is clicked, it loads a list of possible
parent projects belonging to the ecosystem.
When the form is saved successfully, it redirects to
the new project's route.

Signed-off-by: Eva Millán <evamillan@bitergia.com>