[ui] Create view to add a project
Creates a 'NewProject' view that includes the
'ProjectForm' component. The route of the view
includes the ecosystem id under which the project
will be created to pass it to the form. If the
ecosystem doesn't exist, it displays an error.
Adds a link to create a project under each
ecosystem in the main app.

Signed-off-by: Eva Millán <evamillan@bitergia.com>