[ui] Search projects
Adds the 'Search' component to the sidebar and creates
a new route to display the 'SearchResults' view at
'/search*query*'.
If the search is performed from the sidebar, it navigates
to the search route and passes the filters as query parameters,
eg. '/search?term=example'.
The search results view shows a search box with a filter selector
and a paginated list of projects, using 'ProjectEntry', that match the
search criteria, retrieved with the 'getProjects' query.

Signed-off-by: Eva Millán <evamillan@bitergia.com>