[ui] Drag and drop projects to change their parent
Projects on 'EcosystemTree' can be moved using drag and drop
to change their parent. Valid parent projects are highlighted
when the project is dragged over them.

A project can be moved to another in these cases:
- Both projects belong to the same ecosystem
- The new parent is not already the project's parent.
- The new parent is not a child of the project.
- If the projects are not root projects, they need to have
  the same root.

The functions to check these cases are found on /utils.
Before the move is performed, the user is prompted to confirm
the action with a dialog.

Signed-off-by: Eva Millán <evamillan@bitergia.com>