[ui] Move projects on app sidebar
Adds the 'moveProject' function to 'App' and
passes it to 'EcosystemTree'. The method uses
a GraphQL mutation to change a project's parent.
When the move is successful, the projects on the
sidebar are reloaded, and if the the current view
is 'Project', it reloads the project's information.
A snackbar shows the success and error messages.

Signed-off-by: Eva Millán <evamillan@bitergia.com>