[ui] Add DatasourceList
Creates the DatasourceList component and adds it
to the Project view. The component groups the
datasources by their URI, eg. a GitHub repository,
into a single line and shows all the categories next
to it.
The styles for all the lists are moved to a SCSS
file to be shared between different components.

Signed-off-by: Eva Millán <evamillan@bitergia.com>