[server] Move app files to their own directory
With this change, the files needed to deploy the Bestiary
application (WSGI, URLs, GraphQL schema) are all under
the same directory, so it's easier to distribute the app
and for web servers to use it.

Configuration files are still on their own directory because
these are only available as examples or for developers.
To deploy in production, users should use their own
configuration.

This commit is based on changes in SortingHat

Signed-off-by: JJMerchante <jj.merchante@gmail.com>